doc(socketio): add doc on dynamic namespace
