+++ socketioxide/src/extract/ns_params/mod.rs
@@ -2,10 +2,10 @@ use std::{fmt, sync::Arc};
 
 use de::NsParamDeserializationError;
 use serde::de::DeserializeOwned;
-use smallvec::SmallVec;
 
 use crate::{
     adapter::Adapter,
+    handler::connect::NsParamBuff,
     handler::{ConnectMiddleware, FromConnectParts},
     socket::Socket,
 };
@@ -13,27 +13,33 @@ use crate::{
 mod de;
 #[cfg(feature = "extensions")]
 pub use ns_param_ext::*;
-/// A buffer that holds the namespace parameters.
-/// It should not be used directly, use the [`NsParam`] extractor instead.
-#[derive(Debug, Default)]
-pub struct NsParamBuff<'a>(SmallVec<[(Box<str>, &'a str); 3]>);
-impl<'k, 'v> From<matchit::Params<'k, 'v>> for NsParamBuff<'v> {
-    fn from(params: matchit::Params<'k, 'v>) -> Self {
-        let mut vec = SmallVec::new();
-        for (k, v) in params.iter() {
-            vec.push((k.to_string().into_boxed_str(), v));
-        }
-        Self(vec)
-    }
-}
-impl<'a, 'v> IntoIterator for &'a NsParamBuff<'v> {
-    type Item = &'a (Box<str>, &'v str);
-    type IntoIter = std::slice::Iter<'a, (Box<str>, &'v str)>;
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter()
-    }
-}
 
+/// An Extractor that deserialize the namespace path parameters of the namespace to the provided type.
+///
+/// This extractor only works for the [`ConnectHandler`](crate::handler::ConnectHandler)
+/// and [`ConnectMiddleware`](crate::handler::ConnectMiddleware).
+/// If you want to keep path parameters for later usage you can use the [`KeepNsParam`] middleware helper with the `extensions` feature enabled.
+///
+/// # Errors
+///
+/// In case of deserialization error, if the tracing feature is enabled, an error log will be emitted.
+///
+/// # Example
+/// ```
+/// # use socketioxide::{SocketIo, extract::{NsParam, SocketRef}};
+/// #[derive(Debug, serde::Deserialize)]
+/// struct Params {
+///     id: String,
+///     user_id: String
+/// }
+///
+/// fn handler(socket: SocketRef, NsParam(params): NsParam<Params>) {
+///     println!("new socket with params: {:?}", params);
+/// }
+/// let (_svc, io) = SocketIo::new_svc();
+///
+/// io.ns("/{id}/user/{user_id}", handler).unwrap();
+/// ```
 #[derive(Debug, Clone)]
 pub struct NsParam<T>(pub T);
 impl<A: Adapter, T: DeserializeOwned> FromConnectParts<A> for NsParam<T> {
@@ -49,12 +55,40 @@ impl<A: Adapter, T: DeserializeOwned> FromConnectParts<A> for NsParam<T> {
 }
 
 #[cfg(feature = "extensions")]
+#[cfg_attr(docsrs, doc(cfg(feature = "extensions")))]
 mod ns_param_ext {
     use super::*;
     /// A middleware helper that save the namespace parameters in the socket extensions.
+    /// You can the use then [`Extension`](crate::extract::Extension) extractor to get the deserialized ns parameters.
+    ///
+    /// # Errors
+    ///
+    /// In case of deserialization error, if the tracing feature is enabled, an error log will be emitted.
+    ///
+    /// # Example
+    /// ```
+    /// # use socketioxide::{SocketIo, extract::{Extension, KeepNsParam, SocketRef}, handler::ConnectHandler};
+    /// #[derive(Debug, Clone, serde::Deserialize)]
+    /// struct Params {
+    ///     id: String,
+    ///     user_id: String
+    /// }
+    ///
+    /// fn handler(s: SocketRef) {
+    ///     s.on("event", |Extension(params): Extension<Params>| {
+    ///         println!("socket initialized with {:?} received new event", params);
+    ///     })
+    /// }
+    /// let (_svc, io) = SocketIo::new_svc();
+    ///
+    /// io.ns("/{id}/user/{user_id}", handler.with(KeepNsParam::<Params>::new())).unwrap();
+    /// ```
     #[derive(Debug, Default)]
-    pub struct KeepNsParam<T>(std::marker::PhantomData<T>);
+    pub struct KeepNsParam<T: DeserializeOwned + Clone + Send + Sync + 'static>(
+        std::marker::PhantomData<T>,
+    );
     impl<T: DeserializeOwned + Clone + Send + Sync + 'static> KeepNsParam<T> {
+        /// Create a new [`KeepNsParam`] middleware helper
         pub fn new() -> Self {
             Self(std::marker::PhantomData)
         }
